Move upload_group_configuration to configuration.py
