+++ framework/wazuh/cluster/dapi/requests_list.py
@@ -149,7 +149,7 @@ functions = {
         'type': 'local_master'
     },
     'POST/agents/groups/:group_id/configuration': {
-        'function': Agent.upload_group_configuration,
+        'function': configuration.upload_group_configuration,
         'type': 'local_master'
     },
     'DELETE/agents/groups/:group_id': {
